refactor: Move PartitionFunctionSpecs out of PlanNode.h
PartitionFunction and PartitionFunctionSpec are core concepts used in
various places like exec and connectors. To make exec PlanBuilder test
utilility class connector agnostic, this commit moves them out of PlanNode.h
to a standalone file in the same folder, making it a first class interface
in velox/core.
