+++ velox/core/PlanNode.h
@@ -22,6 +22,7 @@
 #include "velox/common/Enums.h"
 #include "velox/connectors/Connector.h"
 #include "velox/core/Expressions.h"
+#include "velox/core/PartitionFunction.h"
 #include "velox/core/QueryConfig.h"
 #include "velox/vector/VectorStream.h"
 
@@ -2340,63 +2341,6 @@ class LocalMergeNode : public PlanNode {
 };
 
 using LocalMergeNodePtr = std::shared_ptr<const LocalMergeNode>;
-
-/// Calculates partition number for each row of the specified vector.
-class PartitionFunction {
- public:
-  virtual ~PartitionFunction() = default;
-
-  /// @param input RowVector to split into partitions.
-  /// @param [out] partitions Computed partition numbers for each row in
-  /// 'input'.
-  /// @return Returns partition number in case all rows of 'input' are
-  /// assigned to the same partition. In this case 'partitions' vector is left
-  /// unchanged. Used to optimize round-robin partitioning in local exchange.
-  virtual std::optional<uint32_t> partition(
-      const RowVector& input,
-      std::vector<uint32_t>& partitions) = 0;
-};
-
-/// Factory class for creating PartitionFunction instances.
-class PartitionFunctionSpec : public ISerializable {
- public:
-  /// If 'localExchange' is true, the partition function is used for local
-  /// exchange within a velox task.
-  virtual std::unique_ptr<PartitionFunction> create(
-      int numPartitions,
-      bool localExchange = false) const = 0;
-
-  virtual ~PartitionFunctionSpec() = default;
-
-  virtual std::string toString() const = 0;
-};
-
-using PartitionFunctionSpecPtr = std::shared_ptr<const PartitionFunctionSpec>;
-
-class GatherPartitionFunctionSpec : public PartitionFunctionSpec {
- public:
-  std::unique_ptr<PartitionFunction> create(
-      int /*numPartitions*/,
-      bool /*localExchange*/) const override {
-    VELOX_UNREACHABLE();
-  }
-
-  std::string toString() const override {
-    return "gather";
-  }
-
-  folly::dynamic serialize() const override {
-    folly::dynamic obj = folly::dynamic::object;
-    obj["name"] = "GatherPartitionFunctionSpec";
-    return obj;
-  }
-
-  static PartitionFunctionSpecPtr deserialize(
-      const folly::dynamic& /* obj */,
-      void* /* context */) {
-    return std::make_shared<GatherPartitionFunctionSpec>();
-  }
-};
 
 /// Partitions data using specified partition function. The number of
 /// partitions is determined by the parallelism of the upstream pipeline. Can
